refactored into routes for implementation with the backend, created necessaru screens based on who is accessing
