+++ src/client/app.css
@@ -862,3 +862,167 @@ input {
   font-weight: 700;
   cursor: pointer;
 }
+
+.routePage {
+  min-height: calc(100vh - 80px);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.routeHero,
+.routeCard {
+  width: 100%;
+  max-width: 760px;
+  background: white;
+  color: #111827;
+  border-radius: 28px;
+  padding: 36px;
+  box-shadow: 0 30px 80px rgba(0, 0, 0, 0.25);
+}
+
+.routeHero {
+  text-align: center;
+}
+
+.routeBadge {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  padding: 8px 14px;
+  border-radius: 999px;
+  background: #f5edff;
+  color: #7000e0;
+  font-weight: 800;
+  font-size: 12px;
+  letter-spacing: 0.04em;
+  text-transform: uppercase;
+}
+
+.routeTitle {
+  margin: 18px 0 12px;
+  font-size: clamp(32px, 6vw, 56px);
+  line-height: 1.05;
+  color: #303062;
+}
+
+.routeIntro {
+  max-width: 620px;
+  margin: 0 auto;
+  font-size: 16px;
+  line-height: 1.7;
+  color: #6b7280;
+}
+
+.routeSectionTitle {
+  margin: 14px 0 10px;
+  font-size: 32px;
+  line-height: 1.1;
+  color: #303062;
+}
+
+.routeSectionText {
+  margin: 0;
+  font-size: 15px;
+  line-height: 1.7;
+  color: #6b7280;
+}
+
+.routeActions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 12px;
+  margin-top: 28px;
+}
+
+.routeBtn {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 46px;
+  padding: 0 18px;
+  border-radius: 14px;
+  border: 1px solid rgba(17, 24, 39, 0.08);
+  background: #f8f9fd;
+  color: #111827;
+  text-decoration: none;
+  font-weight: 700;
+  transition: transform 0.12s ease, box-shadow 0.12s ease,
+    background 0.12s ease;
+}
+
+.routeBtn:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 10px 24px rgba(0, 0, 0, 0.08);
+}
+
+.routeBtnPrimary {
+  background: #7000e0;
+  border-color: #7000e0;
+  color: white;
+}
+
+.studyHeader {
+  width: 100%;
+  margin-bottom: 22px;
+}
+
+.studyHeaderInner {
+  display: flex;
+  gap: 10px;
+  justify-content: center;
+  flex-wrap: wrap;
+}
+
+.studyNavLink {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 42px;
+  padding: 0 16px;
+  border-radius: 12px;
+  background: rgba(255, 255, 255, 0.14);
+  color: white;
+  text-decoration: none;
+  border: 1px solid rgba(255, 255, 255, 0.18);
+  font-weight: 700;
+}
+
+.studyNavLink:hover {
+  background: rgba(255, 255, 255, 0.2);
+}
+
+.budgetWarning {
+  margin-top: 16px;
+  padding: 12px 14px;
+  border-radius: 14px;
+  background: rgba(255, 255, 255, 0.92);
+  color: #9c5a14;
+  font-weight: 700;
+}
+
+@media (max-width: 980px) {
+  body {
+    padding: 20px;
+  }
+
+  .simulatorLayout {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .postPanel,
+  .phone {
+    width: 100%;
+    max-width: 460px;
+  }
+
+  .envelopeShell {
+    width: 100%;
+  }
+
+  .routeHero,
+  .routeCard {
+    padding: 28px 22px;
+  }
+}
